requirements.txt
aiohttp==3.8.6
pydantic==2.4.2
pydantic_settings==2.0.3
python-dotenv==1.0.0
